N0030_AFC_AI_CD/trunk/Source/afcAIAnalysis/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
ibm-db==2.0.7
idna==2.6
lxml==4.0.0
numpy==1.13.1
psycopg2==2.7.3.1
PyInstaller==3.2.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
scikit-learn==0.19.0
scipy==0.19.1
six==1.11.0
tsdbpy==1.0.0
urllib3==1.22
